requirements.txt
SpeechRecognition==3.10.4
pyaudio==0.2.14
setuptools==80.9.0
Pillow==10.1.0
psutil==5.9.8
PyAutoGUI==0.9.54
pyinstaller==6.3.0
